✅ test: the handoff's guards hold what its spec now promises
Copilot's review of the design-system spec found three places where the
spec promised more than the tests held. Each one was measured before it
was fixed.

- A counted figure was cast to int before it was compared, so "56.9"
  passed as the count 56. Measured both ways: the old comparison passes
  a page showing 56.9, the new one fails it naming the text and the
  SDK's count, and the committed 56 passes.
- A null probe was skipped with no reason, and that is how the APCA
  palette stayed a request after this pull request shipped it. A null
  probe now needs a "checkedByHand" reason. The stricter test failed
  naming the five items before status.json moved. The detents and the
  dragged state got probes that are absent from the public API today,
  semantics parity and live blocks say why no probe can prove them, and
  the APCA palette is shipped.
- The APCA requirement claimed Lc 15 for BorderStrong against all three
  grounds, while the handoff and the test measure Surface and
  Background. The spec now says what is held. BorderStrong measures
  Lc 13.7 on SurfaceSubtle in the dark palette, and whether a field sits
  there is a design question, #436.

+++ tests/eQuantic.UI.Native.Engine.Tests/HandoffFigureTests.cs
@@ -93,7 +93,8 @@ public class HandoffFigureTests
             var text = match.Groups["text"].Value;
             if (!counted.TryGetValue(name, out var actual))
                 offences.Add($"{where}: data-figure=\"{name}\" is not a figure this test knows how to count");
-            else if (!TryShown(text, out var shown) || (int)shown != actual)
+            // Compared as read, never truncated first: a cast let "56.9" pass as the count 56.
+            else if (!TryShown(text, out var shown) || shown != actual)
                 offences.Add($"{where}: the page shows \"{text}\" and the SDK has {actual}");
         }
